lib was rebuilt & demos were refactored

+++ HttpRemoting/ViewController.m
@@ -369,7 +369,7 @@ enum eye_color
 #pragma mark -
 #pragma mark UIAlertViewDelegate Methods 
 
--(void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(int)index {
+-(void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)index {
 	//[alertView release];	
 }
